🧹 chore: reorganize FullURL helper and docs
